refactor SortingQuizQuestion using separate AlgorithmQuizQuestion component
