+++ src/main/algorithms/types.ts
@@ -4,3 +4,8 @@ export interface AlgorithmImplementation<T extends SimulationStep> {
     run: (...args: any) => Simulation<T>,
     description: () => string[],
 }
+
+export type AlgorithmData<A> = {
+    label: string,
+    value: A,
+}
